removed scales from vis
