+++ src/vis/EagerVis.tsx
@@ -283,28 +283,6 @@ export function EagerVis({
     return currMap;
   }, [selected]);
 
-  const scales: Scales = React.useMemo(
-    () => ({
-      color: d3v7
-        .scaleOrdinal()
-        .range(
-          colors || [
-            getCssValue('visyn-c1'),
-            getCssValue('visyn-c2'),
-            getCssValue('visyn-c3'),
-            getCssValue('visyn-c4'),
-            getCssValue('visyn-c5'),
-            getCssValue('visyn-c6'),
-            getCssValue('visyn-c7'),
-            getCssValue('visyn-c8'),
-            getCssValue('visyn-c9'),
-            getCssValue('visyn-c10'),
-          ],
-        ),
-    }),
-    [colors],
-  );
-
   const commonProps = {
     showSidebar,
     setShowSidebar,
@@ -366,7 +344,6 @@ export function EagerVis({
               selectedMap={selectedMap}
               selectedList={selected}
               columns={columns}
-              scales={scales}
               showSidebar={showSidebar}
               showCloseButton={showCloseButton}
               closeButtonCallback={closeCallback}
