+++ src/vis/general/InvalidCols.tsx
@@ -1,9 +1,9 @@
-import { Alert, Center, Stack, rem, Overlay } from '@mantine/core';
+import { Alert, Center, Stack } from '@mantine/core';
 import * as React from 'react';
 
-export function InvalidCols({ headerMessage, bodyMessage }: { headerMessage: string; bodyMessage: string }) {
+export function InvalidCols({ headerMessage, bodyMessage, style }: { headerMessage: string; bodyMessage: string; style: React.CSSProperties }) {
   return (
-    <Stack h="100%">
+    <Stack h="100%" style={style}>
       <Center h="100%">
         <Alert title={headerMessage} color="yellow">
           {bodyMessage}
